+++ include/experimental/mdarray
@@ -24,4 +24,4 @@
   #define MDSPAN_IMPL_PROPOSED_NAMESPACE experimental
 #endif
 
-#include "../mdspan/experimental/mdarray.hpp"
+#include "../mdspan/mdarray.hpp"
